Consolidate replayio and test-utils to use shared cachedFetch method
