+++ packages/replay/src/metadata/source.ts
@@ -1,8 +1,7 @@
+import { cachedFetch } from "@replay-cli/shared/cachedFetch";
 import dbg from "debug";
 import fs from "fs";
-import fetch, { RequestInit } from "node-fetch";
 import { create, defaulted, number, object, optional } from "superstruct";
-
 import { UnstructuredMetadata } from "../types";
 import { envString } from "./env";
 
@@ -25,61 +24,6 @@ class GitHubHttpError extends Error {
 type CacheEntry = { json: any | null; status: number; statusText: string };
 
 export const cache: Map<string, CacheEntry> = new Map();
-
-// Note that this method should not be used for GraphQL queries because it caches responses by URL.
-// TODO [PRO-676] Import this from the "shared" package
-export async function fetchWithCacheAndRetry(
-  url: string,
-  init?: RequestInit,
-  options: {
-    baseDelay?: number;
-    maxAttempts?: number;
-    shouldRetry?: (response: fetch.Response) => Promise<boolean>;
-  } = {}
-): Promise<CacheEntry> {
-  const { baseDelay = 1_000, maxAttempts = 3, shouldRetry: shouldRetryFn } = options;
-
-  let attempt = 0;
-
-  while (!cache.has(url)) {
-    attempt++;
-
-    const resp = await fetch(url, init);
-    if (resp.status === 200) {
-      const json = await resp.json();
-      cache.set(url, {
-        json,
-        status: resp.status,
-        statusText: resp.statusText,
-      });
-    } else if (attempt < maxAttempts) {
-      if (shouldRetryFn) {
-        const shouldRetry = await shouldRetryFn(resp);
-        if (!shouldRetry) {
-          cache.set(url, {
-            json: null,
-            status: resp.status,
-            statusText: resp.statusText,
-          });
-          break;
-        }
-      }
-
-      // Retry with exponential backoff (e.g. 1s, 2s, 4s, ...)
-      const delay = Math.pow(2, attempt) * baseDelay;
-      await new Promise(resolve => setTimeout(resolve, delay));
-    } else {
-      // If we've run out of retries, store and return the error
-      cache.set(url, {
-        json: null,
-        status: resp.status,
-        statusText: resp.statusText,
-      });
-    }
-  }
-
-  return cache.get(url)!;
-}
 
 export function resetCache(url?: string) {
   if (url) {
@@ -177,7 +121,7 @@ async function expandCommitMetadataFromGitHub(repo: string, sha?: string) {
 
   debug("Fetching commit metadata from %s with %d char token", url, GITHUB_TOKEN?.length || 0);
 
-  const resp = await fetchWithCacheAndRetry(url, {
+  const resp = await cachedFetch(url, {
     headers: GITHUB_TOKEN
       ? {
           Authorization: `token ${GITHUB_TOKEN}`,
@@ -223,7 +167,7 @@ async function expandMergeMetadataFromGitHub(repo: string, pr?: string) {
 
   debug("Fetching merge metadata from %s with %d char token", url, GITHUB_TOKEN?.length || 0);
 
-  const resp = await fetchWithCacheAndRetry(url, {
+  const resp = await cachedFetch(url, {
     headers: GITHUB_TOKEN
       ? {
           Authorization: `token ${GITHUB_TOKEN}`,
