+++ packages/replayio/tsconfig.json
@@ -6,7 +6,7 @@
     "resolveJsonModule": true,
     "types": ["node", "jest"]
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "../shared/src/cachedFetch.test.ts", "../shared/src/cachedFetch.ts"],
   "exclude": ["**/*.test.ts"],
   "references": [
     {
